bank: add a route that resets all data
for use in E2E tests
